move src, update a few deps, use jest for tests
