+++ .eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": "buildo",
+  "rules": {
+    "max-len": 0
+  }
+}
