+++ .hophoprc
@@ -1,0 +1,3 @@
+toggl: n
+branchPrefix: n
+branchSuffix: y
