+++ package.json
@@ -1,30 +1,42 @@
 {
-  "name": "state",
+  "name": "buildo-state",
   "version": "0.2.0",
-  "description": "",
-  "main": "index.js",
+  "description": "State container for react apps with provider and connect HOCs",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "mocha"
+    "test": "jest",
+    "lint": "eslint src",
+    "release-version": "smooth-release",
+    "preversion": "npm run lint",
+    "prepublish": "npm run build",
+    "build": "rm -rf lib && mkdir lib && babel src -d lib"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/buildo/state.git"
   },
   "author": "Giovanni Gonzaga <giovanni@buildo.io>",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/buildo/state/issues"
   },
   "homepage": "https://github.com/buildo/state#readme",
   "devDependencies": {
-    "babel": "^5.8.38",
-    "expect": "^1.20.1",
-    "lodash": "^4.11.2",
-    "mocha": "^2.4.5",
+    "babel-cli": "^6.22.2",
+    "babel-eslint": "^7.1.1",
+    "babel-plugin-transform-class-properties": "^6.22.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-es2015-classes": "^6.22.0",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-stage-0": "^6.22.0",
+    "eslint": "^3.14.0",
+    "eslint-config-buildo": "^0.2.0",
+    "jest": "^18.1.0",
     "react": "^0.14.8",
-    "require-dir": "^0.3.0",
     "rxjs": "5.0.0-beta.8",
     "sinon": "^1.17.4",
+    "smooth-release": "^6.0.0",
     "tcomb": "^3.0.0"
   }
 }
