+++ src/shallowEqual.js
